+++ database/factories/Server.php
@@ -1,0 +1,16 @@
+<?php
+
+/** @var \Illuminate\Database\Eloquent\Factory $factory */
+
+use App\Server;
+use Faker\Generator as Faker;
+
+$factory->define(Server::class, function (Faker $faker) {
+    return [
+        'baseUrl' => 'https://test.notld/bigbluebutton/',
+        'salt' => $faker->sha1,
+        'description' => $faker->text,
+        'status' => true,
+        'offline' => false
+    ];
+});