+++ database/migrations/2020_10_03_153503_add_current_stat_to_rooms_table.php
@@ -1,0 +1,28 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddCurrentStatToRoomsTable extends Migration
+{
+    public function up()
+    {
+        Schema::table('rooms', function (Blueprint $table) {
+            $table->integer('participant_count')->nullable();
+            $table->integer('listener_count')->nullable();
+            $table->integer('voice_participant_count')->nullable();
+            $table->integer('video_count')->nullable();
+        });
+    }
+
+    public function down()
+    {
+        Schema::table('rooms', function (Blueprint $table) {
+            $table->dropColumn('participant_count');
+            $table->dropColumn('listener_count');
+            $table->dropColumn('voice_participant_count');
+            $table->dropColumn('video_count');
+        });
+    }
+}