+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a middleware to check whether the request is in sync with model of the database and not stale ([#40], [#41])
 - Added management of users and profile page ([#10], [#66])
 - Added management of application settings ([#55], [#60])
+- Added artisan command to check server and meeting status, build live and historical data ([#64], [#65])
 
 
 [#1]: https://github.com/THM-Health/PILOS/issues/1
@@ -62,7 +63,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#55]: https://github.com/THM-Health/PILOS/issues/55
 [#60]: https://github.com/THM-Health/PILOS/pull/60
 [#62]: https://github.com/THM-Health/PILOS/issues/62
-[#66]: https://github.com/THM-Health/PILOS/pull/66
+[#64]: https://github.com/THM-Health/PILOS/issues/64
+[#65]: https://github.com/THM-Health/PILOS/pull/65
+[#65]: https://github.com/THM-Health/PILOS/pull/66
 [#73]: https://github.com/THM-Health/PILOS/pull/73
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/3c8359cdb0395546fe97aeabf1a40f93002b182c...HEAD